changed crop_with_seconds so it uses integer idx

+++ sound/util.py
@@ -533,7 +533,7 @@ class Sound(object):
         return cropped_sound
 
     def crop_with_seconds(self, first_second, last_second):
-        return self.crop_with_idx(round(first_second * self.sr), round(last_second * self.sr))
+        return self.crop_with_idx(int(round(first_second * self.sr)), int(round(last_second * self.sr)))
 
     def mix_in(self,
                sound,
